feat(gltf_auto_export): cleaned up folder & modules structure, moved a few files around for coherence
