+++ tools/gltf_auto_export/auto_export/internals.py
