+++ tools/gltf_auto_export/auto_export/operators.py
@@ -4,8 +4,8 @@ from bpy.types import Operator
 from bpy_extras.io_utils import ExportHelper
 from bpy.props import (IntProperty, StringProperty)
 from .preferences import (AutoExportGltfAddonPreferences, AutoExportGltfPreferenceNames)
-from ..helpers_scenes import (get_scenes)
-from ..helpers_collections import (get_exportable_collections)
+from ..helpers.helpers_scenes import (get_scenes)
+from ..helpers.helpers_collections import (get_exportable_collections)
 from .auto_export import auto_export
 
 class AutoExportGLTF(Operator, AutoExportGltfAddonPreferences, ExportHelper):
