+++ tools/gltf_auto_export/helpers/helpers_collections.py
