+++ tools/gltf_auto_export/helpers/object_makers.py
