+++ tools/gltf_auto_export/helpers/to_remove_later.py
