+++ tools/gltf_auto_export/ui/main.py
@@ -11,8 +11,8 @@ from bpy.props import (BoolProperty,
 from ..auto_export import auto_export
 
 from ..auto_export.preferences import (AutoExportGltfAddonPreferences, AutoExportGltfPreferenceNames)
-from ..helpers_scenes import (get_scenes)
-from ..helpers_collections import (get_exportable_collections)
+from ..helpers.helpers_scenes import (get_scenes)
+from ..helpers.helpers_collections import (get_exportable_collections)
 ######################################################
 ## ui logic & co
